Move :<>= tests to their own spec
